refractor(api): group questions and answers routes with Express Router
